style: 182-move total-cost-emitter into helpers
