+++ src/components/cart/layout.ts
@@ -1,7 +1,7 @@
 import { createDiv, createSpan } from '../../utils/create-elements/create-tags';
 import { CART } from '../../pages/cart/constants';
 import { createProductsWrapper } from './items-wrapper';
-import { createTotalCostEmitter } from './item/total-cost';
+import { createTotalCostEmitter } from '../../helpers/total-cost-emitter';
 
 export function cartLayout(): HTMLDivElement {
   const layout = createDiv({ classes: CART.wrapper });
