+++ src/utils/render-cart/render-cart.ts
@@ -5,7 +5,7 @@ import { LINK_CONFIG } from '../../shared/ui-config/link';
 import { createCartItem } from '../../components/cart/item/item';
 import { isFetchError } from '../type-guards/is-fetch-error';
 import { CART_MESSAGES } from '../../pages/cart/constants';
-import { costEventEmitter } from '../../components/cart/item/total-cost';
+import { costEventEmitter } from '../../helpers/total-cost-emitter';
 
 export async function renderCartItems(itemsWrapper: HTMLDivElement): Promise<void> {
   itemsWrapper.replaceChildren();
